Disable libc and openssl preloads by default
One or both of these seem to be triggerning an asan failure, which I
haven't tried to further debug. I think though that I remember there
being some issues before with the libc preload in particular
interfering, and google results for the asan failure suggested it could
have to do with dynamically linked libraries that aren't themselves
compiled with asan, so just dropping both.